refactor: reorganise cpu code
