+++ partyboy-core/src/cpu/instructions/opcodes.rs
@@ -1,88 +1,42 @@
-use std::fmt::Debug;
-
-use super::{register::Flag, Cpu};
+use super::super::{register::Flag, Cpu};
+use super::instruction::{Instruction, InstructionState, InstructionStep};
 use crate::bus::Bus;
 
 use paste::paste;
-#[cfg(feature = "serde")]
-use serde::{Deserialize, Serialize};
 
-type InstructionFn = fn(&mut Cpu, &mut Bus) -> InstructionState;
-
-#[derive(Clone, Copy)]
-pub enum InstructionState {
-    /// There are more steps to execute, wait 4T
-    InProgress,
-
-    /// There are more steps to execute, execute the next step instantly
-    ExecNextInstantly,
-
-    /// We've finished fully executing the opcode
-    Finished,
-
-    /// Are we finishing the instruction early?
-    Branch(bool),
-}
-
-pub(crate) enum InstructionStep {
-    Standard(InstructionFn),
-    Instant(InstructionFn),
-}
-
-#[derive(Clone, Copy)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum InstructionOpcode {
-    InterruptServiceRoutine,
-    Unprefixed(u8),
-    Prefixed(u8),
-}
-
-impl Debug for InstructionOpcode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::InterruptServiceRoutine => f.write_str("ISR"),
-            Self::Unprefixed(arg0) => f.write_str(format!("{:#06X}", arg0).as_str()),
-            Self::Prefixed(arg0) => f.write_str(format!("{:#06X}", 0xCB00 + *arg0 as u16).as_str()),
-        }
-    }
-}
-
-pub(super) struct Instruction {
-    pub steps: Vec<InstructionStep>,
-}
-
-const __FETCH_OPERAND8: InstructionStep = InstructionStep::Standard(|cpu, bus| {
+pub(super) const __FETCH_OPERAND8: InstructionStep = InstructionStep::Standard(|cpu, bus| {
     cpu.operand8 = cpu.fetch(bus);
     InstructionState::InProgress
 });
 
-const __FETCH_OPERAND16: InstructionStep = InstructionStep::Standard(|cpu, bus| {
+pub(super) const __FETCH_OPERAND16: InstructionStep = InstructionStep::Standard(|cpu, bus| {
     let hi = cpu.fetch(bus);
     cpu.operand16 = (hi as u16) << 8 | cpu.operand8 as u16;
     InstructionState::InProgress
 });
 
-const FETCH_OP8_EXECNEXTINSTANT: InstructionStep = InstructionStep::Standard(|cpu, bus| {
-    cpu.operand8 = cpu.fetch(bus);
-    InstructionState::ExecNextInstantly
-});
+pub(super) const FETCH_OP8_EXECNEXTINSTANT: InstructionStep =
+    InstructionStep::Standard(|cpu, bus| {
+        cpu.operand8 = cpu.fetch(bus);
+        InstructionState::ExecNextInstantly
+    });
 
-const BLANK_PROGRESS: InstructionStep =
+pub(super) const BLANK_PROGRESS: InstructionStep =
     InstructionStep::Standard(|_, _| InstructionState::InProgress);
 
-const BLANK_PROGRESS_EXEC_NEXT_INSTANT: InstructionStep =
+pub(super) const BLANK_PROGRESS_EXEC_NEXT_INSTANT: InstructionStep =
     InstructionStep::Standard(|_, _| InstructionState::ExecNextInstantly);
 
-const BRANCH_ZERO: InstructionStep =
+pub(super) const BRANCH_ZERO: InstructionStep =
     InstructionStep::Instant(|cpu, _| InstructionState::Branch(cpu.is_flag_set(Flag::Z)));
 
-const BRANCH_NOT_ZERO: InstructionStep =
+pub(super) const BRANCH_NOT_ZERO: InstructionStep =
     InstructionStep::Instant(|cpu, _| InstructionState::Branch(!cpu.is_flag_set(Flag::Z)));
 
-const BRANCH_CARRY: InstructionStep =
+pub(super) const BRANCH_CARRY: InstructionStep =
     InstructionStep::Instant(|cpu, _| InstructionState::Branch(cpu.is_flag_set(Flag::C)));
 
-const BRANCH_NOT_CARRY: InstructionStep =
+pub(super) const BRANCH_NOT_CARRY: InstructionStep =
     InstructionStep::Instant(|cpu, _| InstructionState::Branch(!cpu.is_flag_set(Flag::C)));
 
 macro_rules! instruction {
@@ -829,7 +783,7 @@ macro_rules! unused_opcode {
     };
 }
 
-fn daa() -> Instruction {
+pub(super) fn daa() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             // https://forums.nesdev.com/viewtopic.php?t=15944
@@ -859,7 +813,7 @@ fn daa() -> Instruction {
     }
 }
 
-fn cpl() -> Instruction {
+pub(super) fn cpl() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             r8!(cpu, a) = !r8!(cpu, a);
@@ -870,7 +824,7 @@ fn cpl() -> Instruction {
     }
 }
 
-fn scf() -> Instruction {
+pub(super) fn scf() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             cpu.set_flag(Flag::C);
@@ -881,7 +835,7 @@ fn scf() -> Instruction {
     }
 }
 
-fn ccf() -> Instruction {
+pub(super) fn ccf() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             if cpu.is_flag_set(Flag::C) {
@@ -897,7 +851,7 @@ fn ccf() -> Instruction {
     }
 }
 
-fn rlca() -> Instruction {
+pub(super) fn rlca() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             let carry = (r8!(cpu, a) & 0x80) >> 7;
@@ -913,7 +867,7 @@ fn rlca() -> Instruction {
     }
 }
 
-fn rrca() -> Instruction {
+pub(super) fn rrca() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             let carry = r8!(cpu, a) & 0b00000001 > 0;
@@ -929,7 +883,7 @@ fn rrca() -> Instruction {
     }
 }
 
-fn rla() -> Instruction {
+pub(super) fn rla() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             let is_carry_set = cpu.is_flag_set(Flag::C);
@@ -946,7 +900,7 @@ fn rla() -> Instruction {
     }
 }
 
-fn rra() -> Instruction {
+pub(super) fn rra() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             let carry = if cpu.is_flag_set(Flag::C) {1 << 7} else {0};
@@ -962,7 +916,7 @@ fn rra() -> Instruction {
     }
 }
 
-fn inc_hl() -> Instruction {
+pub(super) fn inc_hl() -> Instruction {
     instruction! {
         InstructionStep::Standard(|cpu, bus| {
             cpu.temp8 = bus.read_u8(cpu.hl.into());
@@ -980,7 +934,7 @@ fn inc_hl() -> Instruction {
     }
 }
 
-fn dec_hl() -> Instruction {
+pub(super) fn dec_hl() -> Instruction {
     instruction! {
         InstructionStep::Standard(|cpu, bus| {
             cpu.temp8 = bus.read_u8(cpu.hl.into());
@@ -998,7 +952,7 @@ fn dec_hl() -> Instruction {
     }
 }
 
-fn ld_hlmem_u8() -> Instruction {
+pub(super) fn ld_hlmem_u8() -> Instruction {
     instruction! {
         fetch8,
         InstructionStep::Standard(|cpu, bus| {
@@ -1008,7 +962,7 @@ fn ld_hlmem_u8() -> Instruction {
     }
 }
 
-fn ld_ff00_u8_a() -> Instruction {
+pub(super) fn ld_ff00_u8_a() -> Instruction {
     instruction! {
         fetch8,
         InstructionStep::Standard(|cpu, bus| {
@@ -1018,7 +972,7 @@ fn ld_ff00_u8_a() -> Instruction {
     }
 }
 
-fn ld_a_ff00_u8() -> Instruction {
+pub(super) fn ld_a_ff00_u8() -> Instruction {
     instruction! {
         fetch8,
         InstructionStep::Standard(|cpu, bus| {
@@ -1028,7 +982,7 @@ fn ld_a_ff00_u8() -> Instruction {
     }
 }
 
-fn ld_ff00_c_a() -> Instruction {
+pub(super) fn ld_ff00_c_a() -> Instruction {
     instruction! {
         InstructionStep::Standard(|cpu, bus| {
             bus.write_u8(0xFF00 + (r8!(cpu, c) as u16), r8!(cpu, a));
@@ -1037,7 +991,7 @@ fn ld_ff00_c_a() -> Instruction {
     }
 }
 
-fn ld_a_ff00_c() -> Instruction {
+pub(super) fn ld_a_ff00_c() -> Instruction {
     instruction! {
         InstructionStep::Standard(|cpu, bus| {
             r8!(cpu, a) = bus.read_u8(0xFF00 + (r8!(cpu, c) as u16));
@@ -1046,7 +1000,7 @@ fn ld_a_ff00_c() -> Instruction {
     }
 }
 
-fn add_sp_i8() -> Instruction {
+pub(super) fn add_sp_i8() -> Instruction {
     instruction! {
         fetch8,
         BLANK_PROGRESS,
@@ -1068,7 +1022,7 @@ fn add_sp_i8() -> Instruction {
     }
 }
 
-fn ld_hl_sp_i8() -> Instruction {
+pub(super) fn ld_hl_sp_i8() -> Instruction {
     instruction! {
         fetch8,
         InstructionStep::Standard(|cpu, _| {
@@ -1089,7 +1043,7 @@ fn ld_hl_sp_i8() -> Instruction {
     }
 }
 
-fn jp_hl() -> Instruction {
+pub(super) fn jp_hl() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, _| {
             cpu.pc = cpu.hl.into();
@@ -1098,7 +1052,7 @@ fn jp_hl() -> Instruction {
     }
 }
 
-fn stop() -> Instruction {
+pub(super) fn stop() -> Instruction {
     let mut steps = Vec::new();
     steps.push(InstructionStep::Instant(|cpu, bus| {
         cpu.switching_speed = true;
@@ -1122,7 +1076,7 @@ fn stop() -> Instruction {
     Instruction { steps }
 }
 
-fn ld_sp_hl() -> Instruction {
+pub(super) fn ld_sp_hl() -> Instruction {
     instruction! {
         InstructionStep::Standard(|cpu, _| {
             cpu.sp = cpu.hl.into();
@@ -1131,7 +1085,7 @@ fn ld_sp_hl() -> Instruction {
     }
 }
 
-fn ld_mem_u16_a() -> Instruction {
+pub(super) fn ld_mem_u16_a() -> Instruction {
     instruction! {
         fetch16,
         InstructionStep::Standard(|cpu, bus| {
@@ -1141,7 +1095,7 @@ fn ld_mem_u16_a() -> Instruction {
     }
 }
 
-fn ld_a_mem_u16() -> Instruction {
+pub(super) fn ld_a_mem_u16() -> Instruction {
     instruction! {
         fetch16,
         InstructionStep::Standard(|cpu, bus| {
@@ -1151,7 +1105,7 @@ fn ld_a_mem_u16() -> Instruction {
     }
 }
 
-fn di() -> Instruction {
+pub(super) fn di() -> Instruction {
     instruction! {
         InstructionStep::Instant(|_, bus| {
             bus.interrupts.disable_master();
@@ -1160,7 +1114,7 @@ fn di() -> Instruction {
     }
 }
 
-fn ei() -> Instruction {
+pub(super) fn ei() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, bus| {
             if !bus.interrupts.is_master_enabled() && !cpu.ei_delay {
@@ -1173,7 +1127,7 @@ fn ei() -> Instruction {
     }
 }
 
-fn halt() -> Instruction {
+pub(super) fn halt() -> Instruction {
     instruction! {
         InstructionStep::Instant(|cpu, bus| {
             if bus.interrupts.is_master_enabled() {
@@ -1199,7 +1153,7 @@ fn halt() -> Instruction {
     }
 }
 
-fn interrupt_service_routine() -> Instruction {
+pub(super) fn interrupt_service_routine() -> Instruction {
     instruction! {
         BLANK_PROGRESS,
         BLANK_PROGRESS,
@@ -1346,7 +1300,7 @@ macro_rules! cb_op_hl_instr {
 }
 
 impl Cpu {
-    fn rlc(&mut self, val: u8) -> u8 {
+    pub(super) fn rlc(&mut self, val: u8) -> u8 {
         self.set_flag_if_cond_else_clear((val & 0x80) != 0, Flag::C);
 
         let carry = (val & 0x80) >> 7;
@@ -1359,7 +1313,7 @@ impl Cpu {
         result
     }
 
-    fn rrc(&mut self, val: u8) -> u8 {
+    pub(super) fn rrc(&mut self, val: u8) -> u8 {
         let carry = val & 0x01;
         let mut result = val >> 1;
 
@@ -1377,7 +1331,7 @@ impl Cpu {
         result
     }
 
-    fn rl(&mut self, val: u8) -> u8 {
+    pub(super) fn rl(&mut self, val: u8) -> u8 {
         let carry = u8::from(self.is_flag_set(Flag::C));
         let result = (val << 1).wrapping_add(carry);
 
@@ -1389,7 +1343,7 @@ impl Cpu {
         result
     }
 
-    fn rr(&mut self, val: u8) -> u8 {
+    pub(super) fn rr(&mut self, val: u8) -> u8 {
         let mut result = val >> 1;
         if self.is_flag_set(Flag::C) {
             result |= 0x80;
@@ -1403,7 +1357,7 @@ impl Cpu {
         result
     }
 
-    fn sla(&mut self, val: u8) -> u8 {
+    pub(super) fn sla(&mut self, val: u8) -> u8 {
         let result = val << 1;
 
         self.set_flag_if_cond_else_clear(val & 0x80 != 0, Flag::C);
@@ -1414,7 +1368,7 @@ impl Cpu {
         result
     }
 
-    fn sra(&mut self, val: u8) -> u8 {
+    pub(super) fn sra(&mut self, val: u8) -> u8 {
         let result = (val & 0x80) | (val >> 1);
 
         self.set_flag_if_cond_else_clear(val & 0x01 != 0, Flag::C);
@@ -1425,7 +1379,7 @@ impl Cpu {
         result
     }
 
-    fn swap(&mut self, val: u8) -> u8 {
+    pub(super) fn swap(&mut self, val: u8) -> u8 {
         let result = ((val & 0x0F) << 4) | ((val & 0xF0) >> 4);
 
         self.handle_z_flag(result);
@@ -1436,7 +1390,7 @@ impl Cpu {
         result
     }
 
-    fn srl(&mut self, val: u8) -> u8 {
+    pub(super) fn srl(&mut self, val: u8) -> u8 {
         let result = val >> 1;
 
         self.set_flag_if_cond_else_clear(val & 0x01 != 0, Flag::C);
@@ -1448,22 +1402,22 @@ impl Cpu {
     }
 
     #[inline(always)]
-    fn set_flag(&mut self, flag: Flag) {
+    pub(super) fn set_flag(&mut self, flag: Flag) {
         self.af.lo |= flag as u8;
     }
 
     #[inline(always)]
-    fn clear_flag(&mut self, flag: Flag) {
+    pub(super) fn clear_flag(&mut self, flag: Flag) {
         self.af.lo &= !(flag as u8);
     }
 
     #[inline(always)]
-    fn is_flag_set(&self, flag: Flag) -> bool {
+    pub(super) fn is_flag_set(&self, flag: Flag) -> bool {
         self.af.lo & (flag as u8) > 0
     }
 
     #[inline(always)]
-    fn set_flag_if_cond_else_clear(&mut self, cond: bool, flag: Flag) {
+    pub(super) fn set_flag_if_cond_else_clear(&mut self, cond: bool, flag: Flag) {
         match cond {
             true => self.af.lo |= flag as u8,
             false => self.af.lo &= !(flag as u8),
@@ -1471,7 +1425,7 @@ impl Cpu {
     }
 
     #[inline(always)]
-    fn handle_z_flag(&mut self, val: u8) {
+    pub(super) fn handle_z_flag(&mut self, val: u8) {
         if val == 0 {
             self.af.lo |= Flag::Z as u8;
         } else {
@@ -1479,595 +1433,103 @@ impl Cpu {
         }
     }
 
-    fn pop_u16_from_stack(&mut self, bus: &Bus) -> u16 {
+    pub(super) fn pop_u16_from_stack(&mut self, bus: &Bus) -> u16 {
         let val = bus.read_u16(self.sp);
         self.sp = self.sp.wrapping_add(2);
         val
     }
 
-    fn pop_u8_from_stack(&mut self, bus: &Bus) -> u8 {
+    pub(super) fn pop_u8_from_stack(&mut self, bus: &Bus) -> u8 {
         let val = bus.read_u8(self.sp);
         self.sp = self.sp.wrapping_add(1);
         val
     }
 
-    fn push_u8_to_stack(&mut self, bus: &mut Bus, val: u8) {
+    pub(super) fn push_u8_to_stack(&mut self, bus: &mut Bus, val: u8) {
         self.sp = self.sp.wrapping_sub(1);
         bus.write_u8(self.sp, val);
     }
 }
 
-pub struct InstructionCache {
-    interrupt_service_routine: Instruction,
-    instructions: [Instruction; 256],
-    cb_instructions: [Instruction; 256],
+pub(super) fn nop() -> Instruction {
+    instruction!(InstructionStep::Instant(|_, _| InstructionState::Finished))
 }
 
-impl Default for InstructionCache {
-    fn default() -> Self {
-        Self::new()
+pub(super) fn ld_u16_sp() -> Instruction {
+    instruction! {
+        fetch16,
+        BLANK_PROGRESS,
+        InstructionStep::Standard(|cpu, bus| {
+            bus.write_u16(cpu.operand16, cpu.sp);
+            InstructionState::Finished
+        })
     }
 }
 
-impl InstructionCache {
-    pub fn new() -> Self {
-        Self {
-            interrupt_service_routine: interrupt_service_routine(),
-            instructions: Self::gen_instructions(),
-            cb_instructions: Self::gen_cb_instructions(),
-        }
-    }
+pub(super) use __adc;
+pub(super) use __add;
+pub(super) use __and;
+pub(super) use __cp;
+pub(super) use __define_branching_op_macro;
+pub(super) use __define_op_macro;
+pub(super) use __jr;
+pub(super) use __or;
+pub(super) use __pop_r16_af_edgecase;
+pub(super) use __read_hl;
+pub(super) use __sbc;
+pub(super) use __sub;
+pub(super) use __xor;
+pub(super) use add_hl_r16;
+pub(super) use bit;
+pub(super) use branch_condition;
+pub(super) use call_cc_u16;
+pub(super) use call_u16;
+pub(super) use cb_op_hl_instr;
+pub(super) use cb_op_instr;
+pub(super) use dec_r16;
+pub(super) use dec_r8;
+pub(super) use enable_master;
+pub(super) use inc_r16;
+pub(super) use inc_r8;
+pub(super) use instruction;
+pub(super) use jp_cc_u16;
+pub(super) use jp_u16;
+pub(super) use ld_a_mem;
+pub(super) use ld_mem_a;
+pub(super) use ld_memhl_r8;
+pub(super) use ld_r16_u16;
+pub(super) use ld_r8_r8;
+pub(super) use ld_r8_u8;
+pub(super) use pop_r16;
+pub(super) use push_r16;
+pub(super) use r8;
+pub(super) use res;
+pub(super) use ret;
+pub(super) use ret_cc;
+pub(super) use rst_yy;
+pub(super) use set;
+pub(super) use unused_opcode;
 
-    fn gen_instructions() -> [Instruction; 256] {
-        [
-            instruction!(InstructionStep::Instant(|_, _| InstructionState::Finished)),
-            ld_r16_u16!(bc),
-            ld_mem_a!(bc),
-            inc_r16!(bc),
-            inc_r8!(bc, hi),
-            dec_r8!(bc, hi),
-            ld_r8_u8!(bc, hi),
-            rlca(),
-            instruction! { // LD (u16), SP
-                fetch16,
-                BLANK_PROGRESS,
-                InstructionStep::Standard(|cpu, bus| {
-                    bus.write_u16(cpu.operand16, cpu.sp);
-                    InstructionState::Finished
-                })
-            },
-            add_hl_r16!(bc),
-            ld_a_mem!(bc),
-            dec_r16!(bc),
-            inc_r8!(bc, lo),
-            dec_r8!(bc, lo),
-            ld_r8_u8!(bc, lo),
-            rrca(),
-            stop(),
-            ld_r16_u16!(de),
-            ld_mem_a!(de),
-            inc_r16!(de),
-            inc_r8!(de, hi),
-            dec_r8!(de, hi),
-            ld_r8_u8!(de, hi),
-            rla(),
-            jr_i8!(),
-            add_hl_r16!(de),
-            ld_a_mem!(de),
-            dec_r16!(de),
-            inc_r8!(de, lo),
-            dec_r8!(de, lo),
-            ld_r8_u8!(de, lo),
-            rra(),
-            jr_i8!(NZ),
-            ld_r16_u16!(hl),
-            ld_mem_a!(hlplus),
-            inc_r16!(hl),
-            inc_r8!(hl, hi),
-            dec_r8!(hl, hi),
-            ld_r8_u8!(hl, hi),
-            daa(),
-            jr_i8!(Z),
-            add_hl_r16!(hl),
-            ld_a_mem!(hlplus),
-            dec_r16!(hl),
-            inc_r8!(hl, lo),
-            dec_r8!(hl, lo),
-            ld_r8_u8!(hl, lo),
-            cpl(),
-            jr_i8!(NC),
-            ld_r16_u16!(sp),
-            ld_mem_a!(hlminus),
-            inc_r16!(sp),
-            inc_hl(),
-            dec_hl(),
-            ld_hlmem_u8(),
-            scf(),
-            jr_i8!(C),
-            add_hl_r16!(sp),
-            ld_a_mem!(hlminus),
-            dec_r16!(sp),
-            inc_r8!(af, hi),
-            dec_r8!(af, hi),
-            ld_r8_u8!(af, hi),
-            ccf(),
-            // ld b, r8
-            ld_r8_r8!(bc, hi <= bc, hi),
-            ld_r8_r8!(bc, hi <= bc, lo),
-            ld_r8_r8!(bc, hi <= de, hi),
-            ld_r8_r8!(bc, hi <= de, lo),
-            ld_r8_r8!(bc, hi <= hl, hi),
-            ld_r8_r8!(bc, hi <= hl, lo),
-            ld_r8_r8!(bc, hi <= HL),
-            ld_r8_r8!(bc, hi <= af, hi),
-            // ld c, r8
-            ld_r8_r8!(bc, lo <= bc, hi),
-            ld_r8_r8!(bc, lo <= bc, lo),
-            ld_r8_r8!(bc, lo <= de, hi),
-            ld_r8_r8!(bc, lo <= de, lo),
-            ld_r8_r8!(bc, lo <= hl, hi),
-            ld_r8_r8!(bc, lo <= hl, lo),
-            ld_r8_r8!(bc, lo <= HL),
-            ld_r8_r8!(bc, lo <= af, hi),
-            // ld d, r8
-            ld_r8_r8!(de, hi <= bc, hi),
-            ld_r8_r8!(de, hi <= bc, lo),
-            ld_r8_r8!(de, hi <= de, hi),
-            ld_r8_r8!(de, hi <= de, lo),
-            ld_r8_r8!(de, hi <= hl, hi),
-            ld_r8_r8!(de, hi <= hl, lo),
-            ld_r8_r8!(de, hi <= HL),
-            ld_r8_r8!(de, hi <= af, hi),
-            // ld e, r8
-            ld_r8_r8!(de, lo <= bc, hi),
-            ld_r8_r8!(de, lo <= bc, lo),
-            ld_r8_r8!(de, lo <= de, hi),
-            ld_r8_r8!(de, lo <= de, lo),
-            ld_r8_r8!(de, lo <= hl, hi),
-            ld_r8_r8!(de, lo <= hl, lo),
-            ld_r8_r8!(de, lo <= HL),
-            ld_r8_r8!(de, lo <= af, hi),
-            // ld h, r8
-            ld_r8_r8!(hl, hi <= bc, hi),
-            ld_r8_r8!(hl, hi <= bc, lo),
-            ld_r8_r8!(hl, hi <= de, hi),
-            ld_r8_r8!(hl, hi <= de, lo),
-            ld_r8_r8!(hl, hi <= hl, hi),
-            ld_r8_r8!(hl, hi <= hl, lo),
-            ld_r8_r8!(hl, hi <= HL),
-            ld_r8_r8!(hl, hi <= af, hi),
-            // ld l, r8
-            ld_r8_r8!(hl, lo <= bc, hi),
-            ld_r8_r8!(hl, lo <= bc, lo),
-            ld_r8_r8!(hl, lo <= de, hi),
-            ld_r8_r8!(hl, lo <= de, lo),
-            ld_r8_r8!(hl, lo <= hl, hi),
-            ld_r8_r8!(hl, lo <= hl, lo),
-            ld_r8_r8!(hl, lo <= HL),
-            ld_r8_r8!(hl, lo <= af, hi),
-            // ld (HL), r8
-            ld_memhl_r8!(HL <= bc, hi),
-            ld_memhl_r8!(HL <= bc, lo),
-            ld_memhl_r8!(HL <= de, hi),
-            ld_memhl_r8!(HL <= de, lo),
-            ld_memhl_r8!(HL <= hl, hi),
-            ld_memhl_r8!(HL <= hl, lo),
-            halt(),
-            ld_memhl_r8!(HL <= af, hi),
-            // ld a, r8
-            ld_r8_r8!(af, hi <= bc, hi),
-            ld_r8_r8!(af, hi <= bc, lo),
-            ld_r8_r8!(af, hi <= de, hi),
-            ld_r8_r8!(af, hi <= de, lo),
-            ld_r8_r8!(af, hi <= hl, hi),
-            ld_r8_r8!(af, hi <= hl, lo),
-            ld_r8_r8!(af, hi <= HL),
-            ld_r8_r8!(af, hi <= af, hi),
-            // add a, r8
-            add_a_r8!(b),
-            add_a_r8!(c),
-            add_a_r8!(d),
-            add_a_r8!(e),
-            add_a_r8!(h),
-            add_a_r8!(l),
-            add_a_r8!(hl),
-            add_a_r8!(a),
-            // adc a, r8
-            adc_a_r8!(b),
-            adc_a_r8!(c),
-            adc_a_r8!(d),
-            adc_a_r8!(e),
-            adc_a_r8!(h),
-            adc_a_r8!(l),
-            adc_a_r8!(hl),
-            adc_a_r8!(a),
-            // sub a, r8
-            sub_a_r8!(b),
-            sub_a_r8!(c),
-            sub_a_r8!(d),
-            sub_a_r8!(e),
-            sub_a_r8!(h),
-            sub_a_r8!(l),
-            sub_a_r8!(hl),
-            sub_a_r8!(a),
-            // sbc a, r8
-            sbc_a_r8!(b),
-            sbc_a_r8!(c),
-            sbc_a_r8!(d),
-            sbc_a_r8!(e),
-            sbc_a_r8!(h),
-            sbc_a_r8!(l),
-            sbc_a_r8!(hl),
-            sbc_a_r8!(a),
-            // and a, r8
-            and_a_r8!(b),
-            and_a_r8!(c),
-            and_a_r8!(d),
-            and_a_r8!(e),
-            and_a_r8!(h),
-            and_a_r8!(l),
-            and_a_r8!(hl),
-            and_a_r8!(a),
-            // xor a, r8
-            xor_a_r8!(b),
-            xor_a_r8!(c),
-            xor_a_r8!(d),
-            xor_a_r8!(e),
-            xor_a_r8!(h),
-            xor_a_r8!(l),
-            xor_a_r8!(hl),
-            xor_a_r8!(a),
-            // or a, r8
-            or_a_r8!(b),
-            or_a_r8!(c),
-            or_a_r8!(d),
-            or_a_r8!(e),
-            or_a_r8!(h),
-            or_a_r8!(l),
-            or_a_r8!(hl),
-            or_a_r8!(a),
-            // cp a, r8
-            cp_a_r8!(b),
-            cp_a_r8!(c),
-            cp_a_r8!(d),
-            cp_a_r8!(e),
-            cp_a_r8!(h),
-            cp_a_r8!(l),
-            cp_a_r8!(hl),
-            cp_a_r8!(a),
-            ret_cc!(NZ),
-            pop_r16!(bc),
-            jp_cc_u16!(NZ),
-            jp_u16!(),
-            call_cc_u16!(NZ),
-            push_r16!(bc),
-            add_a_r8!(u8),
-            rst_yy!(0),
-            ret_cc!(Z),
-            ret!(),
-            jp_cc_u16!(Z),
-            instruction!(InstructionStep::Instant(|_, _| unimplemented!("CB PREFIX"))),
-            call_cc_u16!(Z),
-            call_u16!(),
-            adc_a_r8!(u8),
-            rst_yy!(0x08),
-            ret_cc!(NC),
-            pop_r16!(de),
-            jp_cc_u16!(NC),
-            unused_opcode!("0xD3"),
-            call_cc_u16!(NC),
-            push_r16!(de),
-            sub_a_r8!(u8),
-            rst_yy!(0x10),
-            ret_cc!(C),
-            ret!(i),
-            jp_cc_u16!(C),
-            unused_opcode!("0xDB"),
-            call_cc_u16!(C),
-            unused_opcode!("0xDD"),
-            sbc_a_r8!(u8),
-            rst_yy!(0x18),
-            ld_ff00_u8_a(),
-            pop_r16!(hl),
-            ld_ff00_c_a(),
-            unused_opcode!("0xE3"),
-            unused_opcode!("0xE4"),
-            push_r16!(hl),
-            and_a_r8!(u8),
-            rst_yy!(0x20),
-            add_sp_i8(),
-            jp_hl(),
-            ld_mem_u16_a(),
-            unused_opcode!("0xEB"),
-            unused_opcode!("0xEC"),
-            unused_opcode!("0xED"),
-            xor_a_r8!(u8),
-            rst_yy!(0x28),
-            ld_a_ff00_u8(),
-            pop_r16!(af),
-            ld_a_ff00_c(),
-            di(),
-            unused_opcode!("0xF4"),
-            push_r16!(af),
-            or_a_r8!(u8),
-            rst_yy!(0x30),
-            ld_hl_sp_i8(),
-            ld_sp_hl(),
-            ld_a_mem_u16(),
-            ei(),
-            unused_opcode!("0xFC"),
-            unused_opcode!("0xFD"),
-            cp_a_r8!(u8),
-            rst_yy!(0x38),
-        ]
-    }
+// Export macros generated by other macros
+//
+// Looks like there's some sort of clippy bug where it thinks the "import" is redundant, if
+// they are removed then the instruction_cache module will fail to compile
 
-    fn gen_cb_instructions() -> [Instruction; 256] {
-        [
-            cb_op_instr!(rlc, b),
-            cb_op_instr!(rlc, c),
-            cb_op_instr!(rlc, d),
-            cb_op_instr!(rlc, e),
-            cb_op_instr!(rlc, h),
-            cb_op_instr!(rlc, l),
-            cb_op_hl_instr!(rlc, hl),
-            cb_op_instr!(rlc, a),
-            cb_op_instr!(rrc, b),
-            cb_op_instr!(rrc, c),
-            cb_op_instr!(rrc, d),
-            cb_op_instr!(rrc, e),
-            cb_op_instr!(rrc, h),
-            cb_op_instr!(rrc, l),
-            cb_op_hl_instr!(rrc, hl),
-            cb_op_instr!(rrc, a),
-            cb_op_instr!(rl, b),
-            cb_op_instr!(rl, c),
-            cb_op_instr!(rl, d),
-            cb_op_instr!(rl, e),
-            cb_op_instr!(rl, h),
-            cb_op_instr!(rl, l),
-            cb_op_hl_instr!(rl, hl),
-            cb_op_instr!(rl, a),
-            cb_op_instr!(rr, b),
-            cb_op_instr!(rr, c),
-            cb_op_instr!(rr, d),
-            cb_op_instr!(rr, e),
-            cb_op_instr!(rr, h),
-            cb_op_instr!(rr, l),
-            cb_op_hl_instr!(rr, hl),
-            cb_op_instr!(rr, a),
-            cb_op_instr!(sla, b),
-            cb_op_instr!(sla, c),
-            cb_op_instr!(sla, d),
-            cb_op_instr!(sla, e),
-            cb_op_instr!(sla, h),
-            cb_op_instr!(sla, l),
-            cb_op_hl_instr!(sla, hl),
-            cb_op_instr!(sla, a),
-            cb_op_instr!(sra, b),
-            cb_op_instr!(sra, c),
-            cb_op_instr!(sra, d),
-            cb_op_instr!(sra, e),
-            cb_op_instr!(sra, h),
-            cb_op_instr!(sra, l),
-            cb_op_hl_instr!(sra, hl),
-            cb_op_instr!(sra, a),
-            cb_op_instr!(swap, b),
-            cb_op_instr!(swap, c),
-            cb_op_instr!(swap, d),
-            cb_op_instr!(swap, e),
-            cb_op_instr!(swap, h),
-            cb_op_instr!(swap, l),
-            cb_op_hl_instr!(swap, hl),
-            cb_op_instr!(swap, a),
-            cb_op_instr!(srl, b),
-            cb_op_instr!(srl, c),
-            cb_op_instr!(srl, d),
-            cb_op_instr!(srl, e),
-            cb_op_instr!(srl, h),
-            cb_op_instr!(srl, l),
-            cb_op_hl_instr!(srl, hl),
-            cb_op_instr!(srl, a),
-            bit!(0, b),
-            bit!(0, c),
-            bit!(0, d),
-            bit!(0, e),
-            bit!(0, h),
-            bit!(0, l),
-            bit!(0, hl),
-            bit!(0, a),
-            bit!(1, b),
-            bit!(1, c),
-            bit!(1, d),
-            bit!(1, e),
-            bit!(1, h),
-            bit!(1, l),
-            bit!(1, hl),
-            bit!(1, a),
-            bit!(2, b),
-            bit!(2, c),
-            bit!(2, d),
-            bit!(2, e),
-            bit!(2, h),
-            bit!(2, l),
-            bit!(2, hl),
-            bit!(2, a),
-            bit!(3, b),
-            bit!(3, c),
-            bit!(3, d),
-            bit!(3, e),
-            bit!(3, h),
-            bit!(3, l),
-            bit!(3, hl),
-            bit!(3, a),
-            bit!(4, b),
-            bit!(4, c),
-            bit!(4, d),
-            bit!(4, e),
-            bit!(4, h),
-            bit!(4, l),
-            bit!(4, hl),
-            bit!(4, a),
-            bit!(5, b),
-            bit!(5, c),
-            bit!(5, d),
-            bit!(5, e),
-            bit!(5, h),
-            bit!(5, l),
-            bit!(5, hl),
-            bit!(5, a),
-            bit!(6, b),
-            bit!(6, c),
-            bit!(6, d),
-            bit!(6, e),
-            bit!(6, h),
-            bit!(6, l),
-            bit!(6, hl),
-            bit!(6, a),
-            bit!(7, b),
-            bit!(7, c),
-            bit!(7, d),
-            bit!(7, e),
-            bit!(7, h),
-            bit!(7, l),
-            bit!(7, hl),
-            bit!(7, a),
-            res!(0, b),
-            res!(0, c),
-            res!(0, d),
-            res!(0, e),
-            res!(0, h),
-            res!(0, l),
-            res!(0, hl),
-            res!(0, a),
-            res!(1, b),
-            res!(1, c),
-            res!(1, d),
-            res!(1, e),
-            res!(1, h),
-            res!(1, l),
-            res!(1, hl),
-            res!(1, a),
-            res!(2, b),
-            res!(2, c),
-            res!(2, d),
-            res!(2, e),
-            res!(2, h),
-            res!(2, l),
-            res!(2, hl),
-            res!(2, a),
-            res!(3, b),
-            res!(3, c),
-            res!(3, d),
-            res!(3, e),
-            res!(3, h),
-            res!(3, l),
-            res!(3, hl),
-            res!(3, a),
-            res!(4, b),
-            res!(4, c),
-            res!(4, d),
-            res!(4, e),
-            res!(4, h),
-            res!(4, l),
-            res!(4, hl),
-            res!(4, a),
-            res!(5, b),
-            res!(5, c),
-            res!(5, d),
-            res!(5, e),
-            res!(5, h),
-            res!(5, l),
-            res!(5, hl),
-            res!(5, a),
-            res!(6, b),
-            res!(6, c),
-            res!(6, d),
-            res!(6, e),
-            res!(6, h),
-            res!(6, l),
-            res!(6, hl),
-            res!(6, a),
-            res!(7, b),
-            res!(7, c),
-            res!(7, d),
-            res!(7, e),
-            res!(7, h),
-            res!(7, l),
-            res!(7, hl),
-            res!(7, a),
-            set!(0, b),
-            set!(0, c),
-            set!(0, d),
-            set!(0, e),
-            set!(0, h),
-            set!(0, l),
-            set!(0, hl),
-            set!(0, a),
-            set!(1, b),
-            set!(1, c),
-            set!(1, d),
-            set!(1, e),
-            set!(1, h),
-            set!(1, l),
-            set!(1, hl),
-            set!(1, a),
-            set!(2, b),
-            set!(2, c),
-            set!(2, d),
-            set!(2, e),
-            set!(2, h),
-            set!(2, l),
-            set!(2, hl),
-            set!(2, a),
-            set!(3, b),
-            set!(3, c),
-            set!(3, d),
-            set!(3, e),
-            set!(3, h),
-            set!(3, l),
-            set!(3, hl),
-            set!(3, a),
-            set!(4, b),
-            set!(4, c),
-            set!(4, d),
-            set!(4, e),
-            set!(4, h),
-            set!(4, l),
-            set!(4, hl),
-            set!(4, a),
-            set!(5, b),
-            set!(5, c),
-            set!(5, d),
-            set!(5, e),
-            set!(5, h),
-            set!(5, l),
-            set!(5, hl),
-            set!(5, a),
-            set!(6, b),
-            set!(6, c),
-            set!(6, d),
-            set!(6, e),
-            set!(6, h),
-            set!(6, l),
-            set!(6, hl),
-            set!(6, a),
-            set!(7, b),
-            set!(7, c),
-            set!(7, d),
-            set!(7, e),
-            set!(7, h),
-            set!(7, l),
-            set!(7, hl),
-            set!(7, a),
-        ]
-    }
-
-    pub(super) fn get(&mut self, opcode: InstructionOpcode) -> &mut Instruction {
-        match opcode {
-            InstructionOpcode::Unprefixed(opcode) => &mut self.instructions[opcode as usize],
-            InstructionOpcode::Prefixed(opcode) => &mut self.cb_instructions[opcode as usize],
-            InstructionOpcode::InterruptServiceRoutine => &mut self.interrupt_service_routine,
-        }
-    }
-}
+#[allow(clippy::single_component_path_imports)]
+pub(super) use adc_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use add_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use and_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use cp_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use jr_i8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use or_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use sbc_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use sub_a_r8;
+#[allow(clippy::single_component_path_imports)]
+pub(super) use xor_a_r8;
